+++ lora.h
@@ -3,7 +3,7 @@
 
 #include "net/netdev.h"
 
-typedef void (lora_data_cb_t)(const char *buffer, size_t len, uint8_t *rssi, int8_t *snr);
+typedef void (lora_data_cb_t)(const char *buffer, size_t len, int16_t *rssi, int8_t *snr);
 
 typedef struct {
     uint8_t bandwidth;
